refactor: CrossSearchResult を ResponseSchema として定義し直す
