+++ server/domain/src/search/models.rs
@@ -1,14 +1,7 @@
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
-use crate::{
-    form::{
-        answer::models::{AnswerId, AnswerLabel, FormAnswerContent},
-        comment::models::CommentId,
-        models::{Form, FormLabel},
-    },
-    user::models::User,
-};
+use crate::form::{answer::models::AnswerId, comment::models::CommentId};
 
 #[derive(Serialize, Deserialize, Debug, PartialEq)]
 pub struct Comment {
@@ -16,14 +9,4 @@ pub struct Comment {
     pub id: CommentId,
     pub content: String,
     pub commented_by: Uuid,
-}
-
-#[derive(Serialize, Debug, PartialEq)]
-pub struct CrossSearchResult {
-    pub forms: Vec<Form>,
-    pub users: Vec<User>,
-    pub answers: Vec<FormAnswerContent>,
-    pub label_for_forms: Vec<FormLabel>,
-    pub label_for_answers: Vec<AnswerLabel>,
-    pub comments: Vec<Comment>,
 }
