+++ server/usecase/src/dto.rs
@@ -1,3 +1,4 @@
+use domain::user::models::User;
 use domain::{
     form::{
         answer::models::{AnswerEntry, AnswerLabel, FormAnswerContent},
@@ -22,6 +23,15 @@ pub struct FormDto {
 }
 
 pub struct UserDto {
-    pub user: domain::user::models::User,
+    pub user: User,
     pub discord_user: Option<DiscordUser>,
+}
+
+pub struct CrossSearchDto {
+    pub forms: Vec<Form>,
+    pub users: Vec<User>,
+    pub answers: Vec<FormAnswerContent>,
+    pub label_for_forms: Vec<FormLabel>,
+    pub label_for_answers: Vec<AnswerLabel>,
+    pub comments: Vec<domain::search::models::Comment>,
 }
